feat: glob-based path_allowed + test improvements
Sikong autonomously implemented during the autonomous iteration cycle:
- Rewrote path_allowed to use proper glob-based matching
- Added comprehensive tests for plan_item_key edge cases
- Added submit_plan_group rejection tests
- Added verification_attempts tracking to EngineAgentNodePacket

Observer: fixed one test expectation mismatch (non-ASCII handling)
and added dangerouslyDisableSandbox to code runtime profile.

+++ src/task_run/harness.rs
@@ -82,7 +82,7 @@ pub struct EngineAgentNodePacket {
     pub allow_write: bool,
     pub budget_max_attempts: u32,
     pub execution_attempts: u32,
-    pub verification_attempts: usize,
+    pub verification_attempts: u32,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize)]
